Conflicts:
	engine/engine/Broken Puppet.exp
	engine/engine/Broken Puppet.lib
	engine/engine/engined.exp
	engine/engine/engined.lib

+++ engine/engine/io/iostatus.cpp
@@ -122,8 +122,8 @@ void CIOStatus::TMouse::update(float elapsed, bool centered){
 	::GetCursorPos(&cursor_screen);
 	::ScreenToClient(App.hWnd, &cursor_screen);
 	
-	dx = (cursor_screen.x - prev_x);
-	dy = (cursor_screen.y - prev_y);
+	dx = (float)(cursor_screen.x - prev_x);
+	dy = (float)(cursor_screen.y - prev_y);
 
 	screen_x = cursor_screen.x;
 	screen_y = cursor_screen.y;
